+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.9] - 2016-03-05
+### Fixed
+- Fix Footer custom social icon bug
+- Fix Sharrre url variable name
+
 ## [1.1.8] - 2016-01-03
 ### Added
 - Change log